move singleton test to tests dir
